Melhorias nos exercícios

+++ src/main/java/br/com/wmw/testes/service/PedidoService.java
@@ -21,10 +21,9 @@ public class PedidoService {
 		return 0;
 	}
 
-	//TODO: Adicionar o desconto por estoque
-	//TODO: Se tiver desconto por estoque habilitado e estoque for > 1000, conceder desconto adicional de 5%
-	public Double calcularDesconto(final Double valor, final Integer quantidade, final Integer estoque) {
+	public double calcularDesconto(final double valor, final long quantidade, final long estoque) {
 		double desconto = 0;
+
 		if (AppConfig.isUsaDescontoPorQuantidade()) {
 			if (quantidade >= 100) {
 				desconto = 20D;
